+++ node/node.go
@@ -374,6 +374,7 @@ func (n *Node) startHTTP(endpoint string, apis []rpc.API, modules []string, cors
 	}
 	// register apis and create handler stack
 	srv := rpc.NewServer()
+	srv.SetBanningMethods(n.config.HTTPBanningMethods)
 	err := RegisterApisFromWhitelist(apis, modules, srv, false)
 	if err != nil {
 		return err
